Cloud_Functions_PubSub/process-cart-finalized/requirements.txt
google-cloud-pubsub
google-cloud-firestore
google-cloud-logging